refactor(agent): split self-management shared infrastructure (#744)
