+++ backend/app/features/self_management_shared/__init__.py
@@ -1,0 +1,1 @@
+"""Shared infrastructure for self-management entry points and built-in agents."""
