+++ backend/tests/personal_agents/test_self_management_agents_service.py
@@ -2,16 +2,16 @@ from __future__ import annotations
 
 import pytest
 
-from app.features.agents_shared.actor_context import (
+from app.features.personal_agents.self_management_agents_service import (
+    self_management_agents_service,
+)
+from app.features.self_management_shared.actor_context import (
     SelfManagementActorType,
     build_self_management_actor_context,
 )
-from app.features.agents_shared.tool_gateway import (
+from app.features.self_management_shared.tool_gateway import (
     SelfManagementSurface,
     SelfManagementToolGateway,
-)
-from app.features.personal_agents.self_management_agents_service import (
-    self_management_agents_service,
 )
 from tests.support.utils import create_a2a_agent, create_user
